chatbot/requirements.txt
langchain
python-dotenv
huggingface_hub
transformers